// web/src/app/modules/game-data/model/action.ts
export enum Action {
  CLOSE = 'CLOSE',
  OPEN = 'OPEN',
  ANSWERED = 'ANSWERED',
  SCORE = 'SCORE',
  SILVER = 'SILVER',
  CATEGORY_SHOW = 'CATEGORY_SHOW',
  CATEGORY_HIDE = 'CATEGORY_HIDE',
  ACTIVE_ROUND = 'ACTIVE_ROUND',
  PRESS_BUTTON = 'PRESS_BUTTON',
  RESET_BUTTON = 'RESET_BUTTON',
  SHOW_QUICKEST = 'SHOW_QUICKEST',
  CLOSE_BUTTON = 'CLOSE_BUTTON',
  TOGGLE_WINNER = 'TOGGLE_WINNER',
  GET_CURRENT = 'GET_CURRENT'
}